[Item03] 생성자나 열거 타입으로 싱글턴임을 보장하라
