--- EffectiveJava/Chapter01/ExampleCode/Diamond.java
@@ -1,8 +1,0 @@
-package ExampleCode;
-
-public class Diamond extends Grade{
-
-    public Diamond(String username, int amount) {
-        super(amount, username, Level.DIAMOND.name());
-    };
-}
