--- EffectiveJava/Chapter01/ExampleCode/Gold.java
@@ -1,9 +1,0 @@
-package ExampleCode;
-
-public class Gold extends Grade {
-
-    public Gold(String username, int amount) {
-        super(amount, username, Level.GOLD.name());
-    };
-
-}
