--- EffectiveJava/Chapter01/ExampleCode/Level.java
@@ -1,5 +1,0 @@
-package ExampleCode;
-
-public enum Level {
-    SILVER, GOLD, PLATINUM, DIAMOND
-}
