--- EffectiveJava/Chapter01/ExampleCode/Platinum.java
@@ -1,10 +1,0 @@
-package ExampleCode;
-
-import ExampleCode.Grade;
-
-public class Platinum extends Grade{
-
-    public Platinum(String username, int amount) {
-        super(amount, username, Level.PLATINUM.name());
-    };
-}
