--- EffectiveJava/Chapter01/ExampleCode/Silver.java
@@ -1,8 +1,0 @@
-package ExampleCode;
-
-public class Silver extends Grade {
-
-    public Silver(String username, int amount) {
-        super(amount, username, Level.SILVER.name());
-    };
-}
