+++ EffectiveJava/src/DI/Item01/ExampleCode/Diamond.java
@@ -1,0 +1,8 @@
+package DI.Item01.ExampleCode;
+
+public class Diamond extends Grade{
+
+    public Diamond(String username, int amount) {
+        super(amount, username, Level.DIAMOND.name());
+    };
+}
