+++ EffectiveJava/src/DI/Item01/ExampleCode/Gold.java
@@ -1,0 +1,9 @@
+package DI.Item01.ExampleCode;
+
+public class Gold extends Grade {
+
+    public Gold(String username, int amount) {
+        super(amount, username, Level.GOLD.name());
+    };
+
+}
