+++ EffectiveJava/src/DI/Item01/ExampleCode/Level.java
@@ -1,0 +1,5 @@
+package DI.Item01.ExampleCode;
+
+public enum Level {
+    SILVER, GOLD, PLATINUM, DIAMOND
+}
