+++ EffectiveJava/src/DI/Item01/ExampleCode/Silver.java
@@ -1,0 +1,8 @@
+package DI.Item01.ExampleCode;
+
+public class Silver extends Grade {
+
+    public Silver(String username, int amount) {
+        super(amount, username, Level.SILVER.name());
+    };
+}
