+++ EffectiveJava/src/Item03/DB.java
@@ -1,0 +1,13 @@
+package Item03;
+
+public class DB {
+
+    private DB(){}
+    public static DB getInstance(){
+        return LazyHolder.INSTANCE;
+    }
+
+    public static class LazyHolder {
+        private static final DB INSTANCE = new DB();
+    }
+}
