--- StudyJava/VideoData/.idea/libraries/com_automation_remarks_video_recorder_junit5_2_0.xml
@@ -1,26 +1,0 @@
-<component name="libraryTable">
-  <library name="com.automation-remarks:video-recorder-junit5:2.0" type="repository">
-    <properties maven-id="com.automation-remarks:video-recorder-junit5:2.0" />
-    <CLASSES>
-      <root url="jar://$MAVEN_REPOSITORY$/com/automation-remarks/video-recorder-junit5/2.0/video-recorder-junit5-2.0.jar!/" />
-      <root url="jar://$MAVEN_REPOSITORY$/com/automation-remarks/video-recorder-core/2.0/video-recorder-core-2.0.jar!/" />
-      <root url="jar://$MAVEN_REPOSITORY$/org/aeonbits/owner/owner-java8/1.0.10/owner-java8-1.0.10.jar!/" />
-      <root url="jar://$MAVEN_REPOSITORY$/org/aeonbits/owner/owner/1.0.10/owner-1.0.10.jar!/" />
-      <root url="jar://$MAVEN_REPOSITORY$/org/apache/commons/commons-lang3/3.8.1/commons-lang3-3.8.1.jar!/" />
-      <root url="jar://$MAVEN_REPOSITORY$/org/zeroturnaround/zt-exec/1.10/zt-exec-1.10.jar!/" />
-      <root url="jar://$MAVEN_REPOSITORY$/commons-io/commons-io/1.4/commons-io-1.4.jar!/" />
-      <root url="jar://$MAVEN_REPOSITORY$/org/slf4j/slf4j-api/1.7.2/slf4j-api-1.7.2.jar!/" />
-      <root url="jar://$MAVEN_REPOSITORY$/org/awaitility/awaitility/3.1.6/awaitility-3.1.6.jar!/" />
-      <root url="jar://$MAVEN_REPOSITORY$/org/hamcrest/hamcrest-library/1.3/hamcrest-library-1.3.jar!/" />
-      <root url="jar://$MAVEN_REPOSITORY$/org/hamcrest/hamcrest-core/1.3/hamcrest-core-1.3.jar!/" />
-      <root url="jar://$MAVEN_REPOSITORY$/org/objenesis/objenesis/2.6/objenesis-2.6.jar!/" />
-      <root url="jar://$MAVEN_REPOSITORY$/log4j/log4j/1.2.17/log4j-1.2.17.jar!/" />
-      <root url="jar://$MAVEN_REPOSITORY$/org/junit/jupiter/junit-jupiter-api/5.4.1/junit-jupiter-api-5.4.1.jar!/" />
-      <root url="jar://$MAVEN_REPOSITORY$/org/apiguardian/apiguardian-api/1.0.0/apiguardian-api-1.0.0.jar!/" />
-      <root url="jar://$MAVEN_REPOSITORY$/org/opentest4j/opentest4j/1.1.1/opentest4j-1.1.1.jar!/" />
-      <root url="jar://$MAVEN_REPOSITORY$/org/junit/platform/junit-platform-commons/1.4.1/junit-platform-commons-1.4.1.jar!/" />
-    </CLASSES>
-    <JAVADOC />
-    <SOURCES />
-  </library>
-</component>
