--- StudyJava/VideoData/.idea/libraries/com_bluecatcode_junit_junit_4_10_extended_1_0_4.xml
@@ -1,12 +1,0 @@
-<component name="libraryTable">
-  <library name="com.bluecatcode.junit:junit-4.10-extended:1.0.4" type="repository">
-    <properties maven-id="com.bluecatcode.junit:junit-4.10-extended:1.0.4" />
-    <CLASSES>
-      <root url="jar://$MAVEN_REPOSITORY$/com/bluecatcode/junit/junit-4.10-extended/1.0.4/junit-4.10-extended-1.0.4.jar!/" />
-      <root url="jar://$MAVEN_REPOSITORY$/junit/junit/4.10/junit-4.10.jar!/" />
-      <root url="jar://$MAVEN_REPOSITORY$/org/hamcrest/hamcrest-core/1.1/hamcrest-core-1.1.jar!/" />
-    </CLASSES>
-    <JAVADOC />
-    <SOURCES />
-  </library>
-</component>
