--- StudyJava/VideoData/VideoData.iml
@@ -1,15 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<module type="JAVA_MODULE" version="4">
-  <component name="NewModuleRootManager" inherit-compiler-output="true">
-    <exclude-output />
-    <content url="file://$MODULE_DIR$">
-      <sourceFolder url="file://$MODULE_DIR$/src" isTestSource="false" />
-      <sourceFolder url="file://$MODULE_DIR$/test" isTestSource="true" />
-    </content>
-    <orderEntry type="inheritedJdk" />
-    <orderEntry type="sourceFolder" forTests="false" />
-    <orderEntry type="library" name="com.bluecatcode.junit:junit-4.10-extended:1.0.4" level="project" />
-    <orderEntry type="library" name="org.assertj:assertj-assertions-generator-maven-plugin:1.6.0" level="project" />
-    <orderEntry type="library" name="com.automation-remarks:video-recorder-junit5:2.0" level="project" />
-  </component>
-</module>
